# components/gov-chat-backend/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Government Services API",
      "version": "1.0.0",
      "description": "API documentation for Government Services microservices",
      "contact": {
        "name": "API Support",
        "email": "support@example.com"
      }
    },
    "servers": [
      {
        "url": "/api",
        "description": "API Server"
      }
    ],
...
    "paths": {
...
      "/admin/database/stats": {
        "get": {
          "summary": "Get database statistics",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database statistics retrieved successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
...
      "/admin/database-operations/reindex": {
        "post": {
          "summary": "Reindex database",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database reindexed successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/database-operations/backup": {
        "post": {
          "summary": "Backup database",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database backed up successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
      "/admin/database-operations/optimize": {
        "post": {
          "summary": "Optimize database",
          "tags": [
            "Admin"
          ],
          "security": [
            {
              "bearerAuth": []
            }
          ],
          "responses": {
            "200": {
              "description": "Database optimized successfully"
            },
            "401": {
              "description": "Unauthorized - authentication required"
            },
            "403": {
              "description": "Forbidden - admin access required"
            },
            "500": {
              "description": "Server error"
            }
          }
        }
      },
...
    }
  }